# backend/swagger/v1/swagger.yaml
openapi: 3.0.1
info:
  title: YSpotify
  version: '1.0'
paths:
...
  '/api/users/{id}':
    get:
      tags:
        - User
      summary: Get a user by id
      parameters:
        - name: id
          in: path
          description: The id of the user
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/User'
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            text/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags:
        - User
      summary: Delete a user
      parameters:
        - name: id
          in: path
          description: The id of the user
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Returns the Deleted User
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/User'
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            text/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: If the User is not authorized to delete the Account (if the User is another one)
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: If the User could not be found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/users/login:
    post:
      tags:
        - User
      summary: Authenticate as a user
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
            encoding:
              username:
                style: form
              password:
                style: form
      responses:
        '200':
          description: Returns the JWT for authentication with this Account
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: string
            text/json:
              schema:
                type: string
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '400':
          description: If the User could not be identified with the given Credentials
  /api/users/register:
    put:
      tags:
        - User
      summary: Register a user with a Given Password and unique Username
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
            encoding:
              username:
                style: form
              password:
                style: form
      responses:
        '201':
          description: The User was Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/User'
            application/json:
              schema:
                $ref: '#/components/schemas/User'
            text/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: If the Username is already taken or the Registration otherwise failed
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/users/{id}/personality':
    get:
      tags:
        - User
      summary: Get a user's "personality" by id
      parameters:
        - name: id
          in: path
          description: The id of the user
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The User was Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/Personality'
            application/json:
              schema:
                $ref: '#/components/schemas/Personality'
            text/json:
              schema:
                $ref: '#/components/schemas/Personality'
        '400':
          description: If the Username is already taken or the Registration otherwise failed
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
...
    Personality:
      type: object
      properties:
        likesDance:
          type: number
          format: float
        tempo:
          type: number
          format: float
        preferInstrumentalOverVocal:
          type: boolean
        valence:
          type: number
          format: float
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    User:
      required:
        - password
        - username
      type: object
      properties:
        id:
          type: integer
          description: Identication of the User
          format: int32
        username:
          minLength: 1
          type: string
          description: Unique Username of the User
        password:
          minLength: 1
          type: string
          description: Hashed Password of the User
        group_name:
          type: string
          description: Secondary Key of the Group the User belongs to
          nullable: true
      additionalProperties: false
      description: "Model of a User of the Service, identified by unsigned integer Id,\r\nContains login credentials as well as a reference to the Group the user belongs to.\r\nThe Model contains the spotify user Id, if the User is linked to a Spotify User."
  securitySchemes:
    Bearer:
      type: http
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]